Added donation sound to settings
In order to facilitate playing MP3s instead of just WAVs,
went from QSound to QMediaplayer. Also, to allow for the
user to have some flexibility, added in the ability to
change the sound file to the settings.

I think with the way I've added in the loadsound function
to a few different places, it SHOULD update if the user changes the
sound after a load. It's the same as the changes I made for the
image, so it *should* work.