Properly handle Top Donor for >100 Donors
I'd created ticket #32 and titled it
"Handle >100 Donors". Really, it was the fact that I'd realized
that without going to the "next page" on the JSON API, I wouldn't
be able to properly handle the top donor. Nothing else I do (as
far as I can tell at the moment) actually depends on getting
all the donors. The other donor-centic files I'm producing
involve the last 5 and last donor. When I was exploring the
JSON API today, I realized that I could actually have the API
deliver the top donor to me. (And also handle the cast when
a donor donates more than once) So, that's a lot easier
than what I'd done here previously.