Adding a user-agent string
I guess this is a result of the DDOS attack that Extra-life.org
suffered on game day, but it started failing grabbing the JSON
data. At first I thought it was because I was grabbing http
instead of https, but nope. It turns out that now I need
to add a user-agent string. Made up my own and that seems
to be just fine to get through.

The amount of places I had to change this just drives home
that I need to do the URLOPEN refactor I've been thinking
about for a while. Just didn't want to make drastic changes
until after game day.