URL load error improvements
In preparation for refactor, improved the error when a URL won't
load to help the user understand which URL gave an error.

Also, in light of the recent issue where this program didn't
work without headers, asks the user to try in their browser
to see if the site is actually offline or if it's another
issue like the header one.