reduced complexity of last5donors
This is why linting can be such a boon. I had created a complex
series of if and elifs when really I only needed an if and an else.
It was the linter telling me this method was too complex coupled
with tackling it next for unittests that made me change it to be
less complex.