Created a ParticipantConf class.
Basically, I noticed this was another refactor target
that could be solved by creating a class. This way we're just
returning fields that are gathered in one place. It also
makes it easier to have a reload_JSON method for after the user
changes something. I think I didn't have that working as well as
it could before someone would update the participant.conf in the
GUI. So I think I now have the scaffolding in place to to handle that.