Continuing for issue #28
Migrated Participant Class to the new ParticipantConf class.
Also, tested that the error will happen if there is a version
mismatch. I don't think I'm going to raise an error because
on the commandline, the person should see that on the screen
when they are running it and an error happens.

And for the GUI I'm going to try and handle it with a popup.