Should now alert the user if the participant conf is out of date
This should close issue #28. It will pop up a setting for a GUI
user that will let them know the participant.conf is out of
date and ask them to replace it with the backup config or
try and just add in the new fields when they save their settings.