Fixed timout issue AND starting on Donor/Donation
So, a timeout issue has been HOUNDING ME. Turns out
it's because I didn't have www in front of the URL
for the team participant.

As for the Donor/Donation refactor. This is for ticket #35.
Realized as I was working on Team stuff that there's a difference
between a donor and a donation. Namely a donor can make many
donations.