GUI QOL changes
First of all, moved the folder check to within the Qtimer to
make sure that if the config is changed, the GUI knows where
to get the folders from.

Second, changed the logic in the Team part of the GUI
because before the commandline output would be very
noisy if the user wasn't part of a team and that would
prevent them from seeing any messages I was printing to the screen.