changes to deal with unavailable URL
Ever since the 2019 DDOS attacks, I've been dealing with issues
where I suddenly can't get to one of the URLs. So I've done a bit
of tuning here to try and make the program resilient against not
being able to access any particular URL. With a future refactor
it may become irrelevant as it may result in fewer calls to the
API. At any rate, I hope to make it not cause issues if the URL
is blocked for some reason.