Final step to making team.py resilient
I think this is the final area where things would fail if it couldn't
get to the URL because of the new DDOS code.

Of course, I realized recently I shouldn't be returning JSON OR a 0.
Instead I should have it raise a custom exception and deal with that
in team.py. For now, however, I think I'm OK with this hack.