Some changes for tests
Changed IPC module to be lowercase. Was still causing issues with
testing, but turns out I wasn't testing extralifedonations methods
anyway, so decided not to worry about it for now. Testing is now
working. However, as I have the tests currently written, the
directory where you run pytest matters for where the
testOutput folder is. So, for simplicity's sake, have one at
origin and one in tests folder for now. I'll make it more
robust later.