Some PyPi fixes + RTD
Did a test install of my PyPi package and it seems that I missed a
setup requirement (because my pip requirements.txt got crowded with
a bunch of stuff that wasn't needed.

Also, parts of my readthedocs.io build failed and I think it was
because of the same package missing from the docs requirements.