Imports modified
So, it turns out that your sys.path is quite different
when you're programming python as a script to be run
vs as a module that can be imported. So I had to change all
the imports. On the plus side, I'm no longer getting a lot of the
weird errors I was getting before on pytest.