A few versioning changes
First of all, we will be following strict semantic versioning now.

Second, since I wanted the version to be a value in the init, I
have changed the setup.cfg to pull it from there so I only have one
place I need to update. (Well, two - need to update it for documentation
, too)