Lots and Lots of Documentation! ✍
Updating lots of docstrings to better document the API.

Also added a few things to the documentation itself.

Bumping patch number since there will no longer be Linux builds.
Since  Pyinstaller is not a cross-compiler, it is very particular
to the system that runs it. Github CI builds were causing issues
on my Fedora system and on my wife's older Ubuntu, it wouldn't
even run for lack of a certain version of libC.