Participant Class Unit Tests
The issue with the team URLs being messed up woke me up to the potential
problem on such a simple thing (I hadn't even been thinking of unit
tests for class variables). So I went ahead and implemented the same
thins for the Participant class. Actually a good idea since I'm
planning on making the base of the URL something that lives in the
package __ini__.py, so this way I'll know if I screw things up.