:wrench: Penultimate participant.py loop refactors
For issue #102
First of all, the loop itself has been refractored and is a lot cleaner now.
It now only loops if you call the file, triggering the if name == main part.
(leading to issue #103 needing to be tackled for the GUI to work)

Also, realized I had been importing urllib errors incorrectly, so I fixed that.

What's left for this refactor is to finish up the unit tests so that I can make sure I'm hitting corner cases.
(As a bonus, the new way of not using a loop should mean that I can write unit tests so that I can test new donations
without having to make a donation, myself) Of course, unit tests should help me in future refactors or bugfixes to help
prevent regressions.