Participant and Team Image Output
resolves #115

I decided to create an HTML as that is the easiest way to make use of it in OBS. I am storing
the data for donations and donors, but I wasn't sure how someone would want to use it, so I'm not yet outputting it. If someone has ideas, open an issue and I'd be happy to implement it.