:wrench: Adding Properties to team.py
Along the way I discovered a class variable that could be refactored into
a local variable. Also had to fix the Pytest that goes along with team
to adapt to the fact that these properties are not editable. (This is a
great side effect of making the properties! Since they are set from the
API there's no reason for an outside class to be making changes to them.)

Also had to fix docstrings in team_participant.py.