🌥️ Coverage Update Completed
I couldn't figure out the test that was bothering me and I haven't had an answer yet from reddit or the Python discord. But I decided it wasn't worth holding everything up. Issue #143 was about getting to 99% coverage and when I run coverage locally I'm getting 99%. If this is the same on Github's CI, I will close issue #143.